FCN-395 refactor(ROSAHCPWizard): extract review section field paths for step validation
Move review metadata under Steps/Review and add getFieldPathsForWizardStepId for reuse by footer validation.
